Pull ScoreInfo into its own top-level class. Hide TrafficPattern from FleetManager subclasses. Remove unused fields from Simulator.
